+++ bidsme/Modules/MRI/MRI.py
@@ -42,12 +42,9 @@ class MRI(baseModule):
 
     bidsmodalities = _MRI.modalities
 
-    __slots__ = ["manufacturer"]
-
     def __init__(self):
         super().__init__()
         self.resetMetaFields()
-        self.manufacturer = None
 
     def _copy_bidsified(self, directory: str, bidsname: str, ext: str) -> None:
         """
@@ -88,90 +85,6 @@ class MRI(baseModule):
                 logger.warning("{} missing bval file for diffusion recording"
                                .format(self.recIdentity()))
 
-    def setManufacturer(self, line: str) -> bool:
-        """
-        Sets manufacturer accordingly to retrieved key line
-        Returns true if manufacturer changes
-
-        Actual manufacturer:
-            Siemens
-            Phillips
-            Unknown
-
-        Parameters
-        ----------
-        line: str
-            key line used to determine manufacturer
-
-        Returns
-        -------
-        bool:
-            True if manufacturer value changes
-        """
-        if line is None:
-            manufacturer = "Unknown"
-        else:
-            lin = line.lower()
-
-            if "siemens" in lin:
-                manufacturer = "Siemens"
-            elif "philips" in lin:
-                manufacturer = "Philips"
-            else:
-                manufacturer = "Unknown"
-
-        if self.manufacturer is None:
-            # First time initialisation
-            self.manufacturer = manufacturer
-            return True
-
-        if manufacturer == self.manufacturer:
-            return False
-        else:
-            self.manufacturer = manufacturer
-            return True
-
-    def setupMetaFields(self, definitions: dict) -> None:
-        """
-        Setup json fields to values from given dictionary.
-
-        Dictionary must contain key "Unknown", and keys
-        correspondand to each of manufacturer.
-
-        Corresponding values are dictionaries with json
-        metafields names (as defined in MRI metafields)
-        as keys and a tuple of DynamicField name, default value
-        If default value is None, no default is defined.
-
-        Parameters
-        ----------
-        definitions: dict
-            dictionary with metadata fields definitions
-        """
-        if self.manufacturer in definitions:
-            meta = definitions[self.manufacturer]
-        else:
-            meta = None
-        meta_default = definitions["Unknown"]
-
-        for metaFields in (self.metaFields_req,
-                           self.metaFields_rec,
-                           self.metaFields_opt):
-            for mod in metaFields:
-                for key in metaFields[mod]:
-                    if meta is not None:
-                        if key in meta:
-                            metaFields[mod][key]\
-                                    = MetaField(meta[key][0],
-                                                scaling=None,
-                                                default=meta[key][1])
-                            continue
-                    if key in meta_default:
-                        metaFields[mod][key]\
-                                = MetaField(meta_default[key][0],
-                                            scaling=None,
-                                            default=meta_default[key][1])
-
     def resetMetaFields(self) -> None:
         """
         Resets currently defined meta fields dictionaries
@@ -199,25 +112,4 @@ class MRI(baseModule):
                 key: None for key in
                 _MRI.optional_modality[mod]}
 
-    def testMetaFields(self):
-        """
-        Test all metafields values and resets not found ones
-        """
-        for metaFields in (self.metaFields_req,
-                           self.metaFields_rec,
-                           self.metaFields_opt):
-            for mod in metaFields:
-                for key, field in metaFields[mod].items():
-                    if field is None or "<<" in field.name:
-                        continue
-                    res = None
-                    try:
-                        res = self.getDynamicField(field.name,
-                                                   default=field.default,
-                                                   raw=True,
-                                                   cleanup=False)
-                    except Exception:
-                        metaFields[mod][key] = None
-                        pass
-                    if res is None:
-                        metaFields[mod][key] = None
+